requirements.txt
# Wargame Simulation Dependencies

# LLM
openai>=1.0.0

# Data handling
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional: for visualization (Task #10)
# flask>=3.0.0
# flask-socketio>=5.3.0
